feat: aggiunta funzionalità di generazione report con validazione del percorso e miglioramenti nella gestione dei formati
